# perf(runtime): optimize PermissionState::check