@AaronO 2021-04-12
> Cool - seems simple enough.
> 
> Out of curiosity wouldn't `#[inline]` also do the trick?

I've refactored this so it no longer uses a macro, it instead inlines it and calls `Self::fmt_access` inside `debug!()`, which will cause a double alloc of the `access` msg in the prompt path, but only for debug builds which is fine IMO.