@AaronO 2021-04-12
Both the flamegraph and the `deno_common` bench confirm this is optimized:

```
Before (main):
perf_now:            	n = 500000, dt = 0.155s, r = 3225806/s, t = 310ns/op

After (this PR):
perf_now:            	n = 5000000, dt = 0.821s, r = 6090134/s, t = 164ns/op
```